Add error handling to get_chain
