+++ agixt/Chain.py
@@ -48,10 +48,13 @@ def get_chain_responses_file_path(chain_name):
 
 class Chain:
     def get_chain(self, chain_name):
-        file_path = get_chain_file_path(chain_name=chain_name)
-        with open(file_path, "r") as f:
-            chain_data = json.load(f)
-        return chain_data
+        try:
+            file_path = get_chain_file_path(chain_name=chain_name)
+            with open(file_path, "r") as f:
+                chain_data = json.load(f)
+            return chain_data
+        except:
+            return {}
 
     def get_chains(self):
         chains = [
@@ -154,30 +157,6 @@ class Chain:
         with open(file_path, "w") as f:
             json.dump(chain_data, f)
 
-    async def run_chain(self, chain_name, user_input=None):
-        file_path = get_chain_responses_file_path(chain_name=chain_name)
-        chain_data = self.get_chain(chain_name=chain_name)
-        logging.info(f"Running chain '{chain_name}'")
-        responses = {}  # Create a dictionary to hold responses.
-        for step_data in chain_data["steps"]:
-            if "prompt" in step_data and "step" in step_data:
-                logging.info(f"Running step {step_data['step']}")
-                step = {}
-                step_response = await self.run_chain_step(
-                    step=step_data, chain_name=chain_name, user_input=user_input
-                )  # Get the response of the current step.
-                step["response"] = step_response
-                step["agent_name"] = step_data["agent_name"]
-                step["prompt"] = step_data["prompt"]
-                step["prompt_type"] = step_data["prompt_type"]
-                responses[step_data["step"]] = step  # Store the response.
-                logging.info(f"Response: {step_response}")
-                # Write the responses to the json file.
-                dt = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-                with open(file_path, "w") as f:
-                    json.dump(responses, f)
-        return responses
-
     def get_step_response(self, chain_name, step_number="all"):
         file_path = get_chain_responses_file_path(chain_name=chain_name)
         try:
@@ -198,69 +177,3 @@ class Chain:
             return responses
         except:
             return {}
-
-    def get_step_content(self, chain_name, prompt_content, user_input, agent_name):
-        new_prompt_content = {}
-        if isinstance(prompt_content, dict):
-            for arg, value in prompt_content.items():
-                if isinstance(value, str):
-                    if "{user_input}" in value:
-                        value = value.replace("{user_input}", user_input)
-                    if "{agent_name}" in value:
-                        value = value.replace("{agent_name}", agent_name)
-                    if "{STEP" in value:
-                        # Count how many times {STEP is in the value
-                        step_count = value.count("{STEP")
-                        for i in range(step_count):
-                            # Get the step number from value between {STEP and }
-                            new_step_number = int(value.split("{STEP")[1].split("}")[0])
-                            # get the response from the step number
-                            step_response = self.get_step_response(
-                                chain_name=chain_name, step_number=new_step_number
-                            )
-                            # replace the {STEPx} with the response
-                            value = value.replace(
-                                f"{{STEP{new_step_number}}}",
-                                f"{step_response['response']}",
-                            )
-                new_prompt_content[arg] = value
-        elif isinstance(prompt_content, str):
-            new_prompt_content = prompt_content
-            if "{user_input}" in prompt_content:
-                new_prompt_content = new_prompt_content.replace(
-                    "{user_input}", user_input
-                )
-            if "{agent_name}" in new_prompt_content:
-                new_prompt_content = new_prompt_content.replace(
-                    "{agent_name}", agent_name
-                )
-            if "{STEP" in prompt_content:
-                step_count = value.count("{STEP")
-                for i in range(step_count):
-                    # Get the step number from value between {STEP and }
-                    new_step_number = int(
-                        prompt_content.split("{STEP")[1].split("}")[0]
-                    )
-                    # get the response from the step number
-                    step_response = self.get_step_response(
-                        chain_name=chain_name, step_number=new_step_number
-                    )
-                    # replace the {STEPx} with the response
-                    new_prompt_content = prompt_content.replace(
-                        f"{{STEP{new_step_number}}}", f"{step_response['response']}"
-                    )
-            if new_prompt_content == {}:
-                new_prompt_content = prompt_content
-        return new_prompt_content
-
-
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser()
-    parser.add_argument("--chain", type=str, default="")
-    parser.add_argument("--user_input", type=str, default="")
-    args = parser.parse_args()
-    chain_name = args.chain
-    user_input = args.user_input
-    import asyncio
-
-    asyncio.run(Chain().run_chain(chain_name=chain_name, user_input=user_input))
